Pivot from Postgres/JPA to Neo4j/Spring Data Neo4j
Link's design had already settled on "anything can link to anything"
(Note<->Concept, Source<->Source, any combination), which needed a
polymorphic table with no DB-level FK integrity in Postgres. That's
Neo4j's native data model for free -- nodes and typed relationships
between any two labels, no polymorphism workaround required. Source
and Note get rebuilt as graph nodes too, not just Concept + Link, so
the whole app lives in one consistent model.

Rewrote SPEC.md and BUILD_PLAN.md for the new stack and resolved the
follow-on design questions the pivot raised, all in favor of
simplicity: Link is a single generic LINKS_TO relationship type (not
four), Concept's many-to-many with Note collapses into Link entirely,
and Tag keeps a maintained nameLower property with a real uniqueness
constraint (not cut, since cutting it would reintroduce a race
condition rather than simplify anything).

Swapped pom.xml/docker-compose.yml/application.yml to Neo4j +
spring-boot-starter-data-neo4j + neo4j-migrations-spring-boot-starter +
testcontainers-neo4j, and cleared the old JPA-dependent domain,
repository, service, controller, and integration test code back to a
bare skeleton -- verified it compiles and boots against a real local
Neo4j container. Controllers, DTOs, and exception handling were mostly
untouched, since they were already persistence-agnostic; only the two
Response DTOs' from(entity) factory methods needed stripping pending
the entity rebuild in upcoming BUILD_PLAN steps.

+++ src/main/java/com/justinus/api/dto/SourcePatchRequest.java
@@ -20,7 +20,7 @@ public record SourcePatchRequest(
         LocalDate dateStarted,
         LocalDate dateFinished,
         SourceStatus status,
-        @Min(1) @Max(5) Short rating,
+        @Min(1) @Max(5) Integer rating,
         String generalNotes
 ) {
 }
